Dec limit fix to return degrees (#200)
V1.11.12 - Updates
- Changed DEC Limit commands to return degrees instead of steps
- Added support to query each DEC limit individually.

+++ src/Mount.hpp
@@ -308,7 +308,7 @@ class Mount
     void clearDecLimitPosition(bool upper);
 
     // Get the DEC limit positions
-    void getDecLimitPositions(long &lowerLimit, long &upperLimit);
+    void getDecLimitPositions(float &lowerLimit, float &upperLimit);
 
     // Asynchronously parks the mount. Moves to the home position and stops all motors.
     void park();
